backend/requirements.txt
# Core Flask dependencies
Flask==3.0.0
Flask-CORS==4.0.0
Flask-JWT-Extended==4.6.0
Flask-Limiter==3.5.0

# Database and ODM
pymongo==3.12
mongoengine==0.27.0

# Data validation and serialization
marshmallow==3.20.2
flask-marshmallow==0.15.0
marshmallow-mongoengine==0.30.0

# Security and password hashing
bcrypt==4.1.2
python-dotenv==1.0.0

# File handling and utilities
Werkzeug==3.0.1
python-magic==0.4.27

# AI/OpenAI integration
# openai==1.6.1
openai==1.109.1

# Audio processing (optional)
pydub==0.25.1

# Development and testing
pytest==7.4.4
pytest-flask==1.3.0
coverage==7.4.0

# Production server
gunicorn==21.2.0

# Logging and monitoring
python-json-logger==2.0.7

# Email validation
email-validator==2.1.0
